Changed the image upload button in NewBook and the way that the single blog displays images

+++ src/components/book/EditBook.jsx
@@ -44,7 +44,7 @@ export default class NewBook extends Component {
                 series: series
             })
             .then(function(value) {
-                window.location.replace('/about');
+                window.location.replace('/books');
             });
     };
     
@@ -59,10 +59,14 @@ export default class NewBook extends Component {
     deleteBook = event => {
         // Prevent page refresh
         event.preventDefault();
-        axios.delete(bookAPI + '/' + this.state.id).then(function(value) {
-            console.log('THis should work');
-            window.location.replace('/about');
-        });
+        const response = window.confirm(
+            'Are you sure you want to delete this book?'
+        );
+        if (response === true) {
+            axios.delete(bookAPI + '/' + this.state.id).then(function(value) {
+                window.location.replace('/about');
+            });
+        }
     };
 
     // Render transforms the components into DOM node that the 
@@ -148,7 +152,7 @@ export default class NewBook extends Component {
                         </label>
                         <input
                             className="input is-rounded"
-                            type="date"
+                            type="month"
                             name="published"
                             value={published}
                             onChange={this.handleChange}
@@ -182,12 +186,14 @@ export default class NewBook extends Component {
                         </button>
                     </div>
                 </form>
-                <button
-                    className="button is-rounded is-danger"
-                    onClick={this.deleteBook}
-                >
-                    Delete Book
-                </button>
+                <div className="columns is-mobile is-centered">
+                    <button
+                        className="button is-rounded is-danger"
+                        onClick={this.deleteBook}
+                    >
+                        Delete Book
+                    </button>
+                </div>
             </div>
         );
     }
